rename .inl to .h
SPM/Xcode gets confused
